split credential API regarding mandatory and optional operations.
