+++ service-base/src/main/java/org/eclipse/hono/service/credentials/BaseCredentialsService.java
@@ -19,7 +19,6 @@ import java.util.Optional;
 import org.eclipse.hono.client.ClientErrorException;
 import org.eclipse.hono.service.EventBusService;
 import org.eclipse.hono.util.CredentialsConstants;
-import org.eclipse.hono.util.CredentialsObject;
 import org.eclipse.hono.util.CredentialsResult;
 import org.eclipse.hono.util.EventBusMessage;
 import org.eclipse.hono.util.TenantConstants;
@@ -31,11 +30,17 @@ import io.vertx.core.json.JsonObject;
 
 /**
  * Base class for implementing {@link CredentialsService}s.
+ * Only implements the mandatory operations of the API.
+ * See {@link CompleteCredentialsService} and {@link CompleteBaseCredentialsService}
+ * for the optional API operations.
  * <p>
- * In particular, this base class provides support for receiving service invocation request messages
+ * This base class provides support for receiving "Get" request messages
  * via vert.x' event bus and route them to specific methods corresponding to the operation indicated
  * in the message.
- * 
+ *
+ * Note: Subclasses <b>MUST</b> implement {@link #getAll(String, String, Handler)}, as it is used
+ * by the routing mechanism (see {@link #processGetRequest(EventBusMessage)}).
+ *
  * @param <T> The type of configuration class this service supports.
  */
 public abstract class BaseCredentialsService<T> extends EventBusService<T> implements CredentialsService {
@@ -50,13 +55,13 @@ public abstract class BaseCredentialsService<T> extends EventBusService<T> imple
      * <p>
      * This method validates the request parameters against the Credentials API
      * specification before invoking the corresponding {@code CredentialsService} methods.
-     * 
+     *
      * @param request The request message.
      * @return A future indicating the outcome of the service invocation.
      * @throws NullPointerException If the request message is {@code null}.
      */
     @Override
-    public final Future<EventBusMessage> processRequest(final EventBusMessage request) {
+    public Future<EventBusMessage> processRequest(final EventBusMessage request) {
 
         Objects.requireNonNull(request);
 
@@ -65,18 +70,12 @@ public abstract class BaseCredentialsService<T> extends EventBusService<T> imple
         switch (CredentialsConstants.CredentialsAction.from(operation)) {
             case get:
                 return processGetRequest(request);
-            case add:
-                return processAddRequest(request);
-            case update:
-                return processUpdateRequest(request);
-            case remove:
-                return processRemoveRequest(request);
             default:
                 return processCustomCredentialsMessage(request);
         }
     }
 
-    private Future<EventBusMessage> processGetRequest(final EventBusMessage request) {
+    Future<EventBusMessage> processGetRequest(final EventBusMessage request) {
 
         final String tenantId = request.getTenant();
         final JsonObject payload = request.getJsonPayload();
@@ -124,93 +123,6 @@ public abstract class BaseCredentialsService<T> extends EventBusService<T> imple
         }
     }
 
-    private Future<EventBusMessage> processAddRequest(final EventBusMessage request) {
-
-        final String tenantId = request.getTenant();
-        final CredentialsObject payload = Optional.ofNullable(request.getJsonPayload())
-                .map(json -> json.mapTo(CredentialsObject.class)).orElse(null);
-
-        if (tenantId == null || payload == null) {
-            return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST));
-        } else if (payload.isValid()) {
-            final Future<CredentialsResult<JsonObject>> result = Future.future();
-            add(tenantId, JsonObject.mapFrom(payload), result.completer());
-            return result.map(res -> {
-                return request.getResponse(res.getStatus())
-                        .setDeviceId(payload.getDeviceId())
-                        .setCacheDirective(res.getCacheDirective());
-            });
-        } else {
-            return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST));
-        }
-    }
-
-    private Future<EventBusMessage> processUpdateRequest(final EventBusMessage request) {
-
-        final String tenantId = request.getTenant();
-        final CredentialsObject payload = Optional.ofNullable(request.getJsonPayload())
-                .map(json -> json.mapTo(CredentialsObject.class)).orElse(null);
-
-        if (tenantId == null || payload == null) {
-            return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST));
-        } else if (payload.isValid()) {
-            final Future<CredentialsResult<JsonObject>> result = Future.future();
-            update(tenantId, JsonObject.mapFrom(payload), result.completer()); 
-            return result.map(res -> {
-                return request.getResponse(res.getStatus())
-                        .setDeviceId(payload.getDeviceId())
-                        .setCacheDirective(res.getCacheDirective());
-            });
-        } else {
-            return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST));
-        }
-    }
-
-    private Future<EventBusMessage> processRemoveRequest(final EventBusMessage request) {
-
-        final String tenantId = request.getTenant();
-        final JsonObject payload = request.getJsonPayload();
-
-        if (tenantId == null || payload == null) {
-            return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST));
-        } else {
-            final String type = getTypesafeValueForField(String.class, payload, CredentialsConstants.FIELD_TYPE);
-            final String authId = getTypesafeValueForField(String.class, payload, CredentialsConstants.FIELD_AUTH_ID);
-            final String deviceId = getTypesafeValueForField(String.class, payload,
-                    CredentialsConstants.FIELD_PAYLOAD_DEVICE_ID);
-
-            // there exist several valid combinations of parameters
-
-            if (type == null) {
-                log.debug("remove credentials request does not contain mandatory type parameter");
-                return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST));
-            } else if (!type.equals(CredentialsConstants.SPECIFIER_WILDCARD) && authId != null) {
-                // delete a single credentials instance
-                log.debug("removing specific credentials [tenant: {}, type: {}, auth-id: {}]", tenantId, type, authId);
-                final Future<CredentialsResult<JsonObject>> result = Future.future();
-                remove(tenantId, type, authId, result.completer());
-                return result.map(res -> {
-                    return request.getResponse(res.getStatus())
-                            .setCacheDirective(res.getCacheDirective());
-                });
-            } else if (deviceId != null && type.equals(CredentialsConstants.SPECIFIER_WILDCARD)) {
-                // delete all credentials for device
-                log.debug("removing all credentials for device [tenant: {}, device-id: {}]", tenantId, deviceId);
-                final Future<CredentialsResult<JsonObject>> result = Future.future();
-                removeAll(tenantId, deviceId, result.completer());
-                return result.map(res -> {
-                    return request.getResponse(res.getStatus())
-                            .setDeviceId(deviceId)
-                            .setCacheDirective(res.getCacheDirective());
-                });
-            } else {
-                log.debug("remove credentials request contains invalid search criteria [type: {}, device-id: {}, auth-id: {}]",
-                        type, deviceId, authId);
-                return Future.failedFuture(new ClientErrorException(HttpURLConnection.HTTP_BAD_REQUEST));
-            }
-        }
-    }
-
     /**
      * Processes a request for a non-standard operation.
      * <p>
@@ -230,17 +142,6 @@ public abstract class BaseCredentialsService<T> extends EventBusService<T> imple
 
     /**
      * {@inheritDoc}
-     * 
-     * This default implementation simply returns an empty result with status code 501 (Not Implemented).
-     * Subclasses should override this method in order to provide a reasonable implementation.
-     */
-    @Override
-    public void add(final String tenantId, final JsonObject otherKeys, final Handler<AsyncResult<CredentialsResult<JsonObject>>> resultHandler) {
-        handleUnimplementedOperation(resultHandler);
-    }
-
-    /**
-     * {@inheritDoc}
      *
      * This default implementation simply returns an empty result with status code 501 (Not Implemented).
      * Subclasses should override this method in order to provide a reasonable implementation.
@@ -252,7 +153,7 @@ public abstract class BaseCredentialsService<T> extends EventBusService<T> imple
 
     /**
      * {@inheritDoc}
-     * 
+     *
      * This default implementation simply returns an empty result with status code 501 (Not Implemented).
      * Subclasses should override this method in order to provide a reasonable implementation.
      */
@@ -263,52 +164,15 @@ public abstract class BaseCredentialsService<T> extends EventBusService<T> imple
 
     /**
      * {@inheritDoc}
-     * 
-     * This default implementation simply returns an empty result with status code 501 (Not Implemented).
-     * Subclasses should override this method in order to provide a reasonable implementation.
+     *
+     * Subclasses MUST implement this method in order to build a working Credential Service.
      */
-    @Override
-    public void getAll(final String tenantId, final String deviceId, final Handler<AsyncResult<CredentialsResult<JsonObject>>> resultHandler) {
-        handleUnimplementedOperation(resultHandler);
-    }
-
-    /**
-     * {@inheritDoc}
-     * 
-     * This default implementation simply returns an empty result with status code 501 (Not Implemented).
-     * Subclasses should override this method in order to provide a reasonable implementation.
-     */
-    @Override
-    public void update(final String tenantId, final JsonObject otherKeys, final Handler<AsyncResult<CredentialsResult<JsonObject>>> resultHandler) {
-        handleUnimplementedOperation(resultHandler);
-    }
-
-    /**
-     * {@inheritDoc}
-     * 
-     * This default implementation simply returns an empty result with status code 501 (Not Implemented).
-     * Subclasses should override this method in order to provide a reasonable implementation.
-     */
-    @Override
-    public void remove(final String tenantId, final String type, final String authId, final Handler<AsyncResult<CredentialsResult<JsonObject>>> resultHandler) {
-        handleUnimplementedOperation(resultHandler);
-    }
-
-    /**
-     * {@inheritDoc}
-     * 
-     * This default implementation simply returns an empty result with status code 501 (Not Implemented).
-     * Subclasses should override this method in order to provide a reasonable implementation.
-     */
-    @Override
-    public void removeAll(final String tenantId, final String deviceId, final Handler<AsyncResult<CredentialsResult<JsonObject>>> resultHandler) {
-        handleUnimplementedOperation(resultHandler);
-    }
+    public abstract void getAll(String tenantId, String deviceId, Handler<AsyncResult<CredentialsResult<JsonObject>>> resultHandler);
 
     /**
      * Handles an unimplemented operation by failing the given handler
      * with a {@link ClientErrorException} having a <em>501 Not Implemented</em> status code.
-     * 
+     *
      * @param resultHandler The handler.
      */
     protected void handleUnimplementedOperation(final Handler<AsyncResult<CredentialsResult<JsonObject>>> resultHandler) {
